+++ Dockerfile
@@ -4,11 +4,10 @@ WORKDIR /workspace
 COPY go.* .
 RUN go mod download
 
-COPY main.go main.go
-COPY controllers/ controllers/
-COPY pkg/ pkg/
+COPY cmd/ cmd/
+COPY internal/ internal/
 RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 \
-    go build -a -o manager main.go
+    go build -a -o manager cmd/main.go
 
 #--------------------------------------------------------------------------------------------------
