+++ internal/controllers/utils.go
@@ -1,11 +1,11 @@
 package controllers
 
 import (
-	configv1 "github.com/borchero/switchboard/pkg/config/v1"
-	"github.com/borchero/switchboard/pkg/ext"
-	"github.com/borchero/switchboard/pkg/integrations"
-	"github.com/borchero/switchboard/pkg/k8s"
-	"github.com/borchero/switchboard/pkg/switchboard"
+	configv1 "github.com/borchero/switchboard/internal/config/v1"
+	"github.com/borchero/switchboard/internal/ext"
+	"github.com/borchero/switchboard/internal/integrations"
+	"github.com/borchero/switchboard/internal/k8s"
+	"github.com/borchero/switchboard/internal/switchboard"
 	cmmeta "github.com/jetstack/cert-manager/pkg/apis/meta/v1"
 	traefik "github.com/traefik/traefik/v2/pkg/provider/kubernetes/crd/traefik/v1alpha1"
 	"go.uber.org/zap"
