+++ internal/controllers/utils_test.go
@@ -3,8 +3,8 @@ package controllers
 import (
 	"testing"
 
+	configv1 "github.com/borchero/switchboard/internal/config/v1"
 	"github.com/borchero/switchboard/internal/k8tests"
-	configv1 "github.com/borchero/switchboard/pkg/config/v1"
 	"github.com/stretchr/testify/assert"
 )
