+++ internal/integrations/interface.go
