+++ internal/k8s/delete.go
