+++ internal/k8s/delete_test.go
