+++ internal/switchboard/hosts_test.go
